yolo: support edgelab mbv2_yolo

+++ components/modules/algorithm/algo_yolo.hpp
@@ -17,10 +17,10 @@ enum
 
 typedef struct
 {
-    uint8_t x;
-    uint8_t y;
-    uint8_t w;
-    uint8_t h;
+    uint16_t x;
+    uint16_t y;
+    uint16_t w;
+    uint16_t h;
     uint8_t confidence;
     uint8_t target;
 } yolo_t;
